Initial work on Caching middleware.
Not particulalry happy with the way the cache key and vary parameters are
being generated. I'm hoping that the current implementation can be a suitable
low-level approach that handles all the cases, but that we can then add
various KeyGenerator implementations that handle the common cases (somewhat
the way routing works).

The initial storage engine will be ccache, but I need to make it support
variations. Also need to fetch on grace mode and eventually add purging
support. Not sure how best to authenticate that yet.