Allow a constraint to be specified on a route. For now, this is a list of allowed parameter values.
It's a bit confusing what belongs to the base Route and what belongs to the
route configuration. Things on the base Route are general information about
the Route which are useful to _other_ middleware, such as caching and upstream
information. Configured route info is specific to the actual routing of a
request. Just because it makes sense doesn't make it a good implementation.