Refactoring how things are configured.
Ultimate, the goal is to move from a middleware-oriented configuration to
a route-oriented configuration. In other words, instead of configuring the
stats middleware, then the caching middle and then the upstream middleware,
you'll instead configure a route and all its middlewares.

This makes it simpler to see what's going on with a specific route.