Handler is now a func that handles a request. Middleware is a special handler that can opt to pass to the next request.
A route can have a StopHandler or a FlowHandler.  A StopHandler is an actual
handler, while a FlowHandler is just a special middleware.

This is all done because most routes with Handlers will want a StopHandler
and shouldn't need to know anything about next/middlewares (since they'll never
call it). This keeps the API clean, but preserves the possibility of introducing
a special FLowHandler if that's desired.

If both a Stop and FlowHandler is specified, the StopHandler will be executed.